use utils/Bag and Config for validate and testing
